Refactor send result handlers into subdirectory
