+++ components/api/test/Altinn.Notifications.Tests/Notifications.Integrations/Wolverine/EmailSendResultHandlerTests.cs
@@ -5,7 +5,7 @@ using Altinn.Notifications.Core.Enums;
 using Altinn.Notifications.Core.Exceptions;
 using Altinn.Notifications.Core.Models.Notification;
 using Altinn.Notifications.Core.Services.Interfaces;
-using Altinn.Notifications.Integrations.Wolverine;
+using Altinn.Notifications.Integrations.Wolverine.Handlers;
 using Altinn.Notifications.Shared.Commands;
 
 using Microsoft.Extensions.Logging.Abstractions;
